first module
This is one module of Graph Visualization project. In this module we make layout of our website. we created some HTML , CSS and JavaScript file. 
In this module first.html is a main file of HTML . You can simply run this file and show the output. Make sure all file path same as the above created.
In this module you can create your own graph using this website and also you can modify your graph. It is very useful feature. So, Use it and enjoy it.